削除機能の実装 指摘箇所修正

+++ app/controllers/items_controller.rb
@@ -1,6 +1,6 @@
 class ItemsController < ApplicationController
   before_action :authenticate_user!, only: [:new, :create, :edit, :destroy]
-  before_action :move_to_index, only: [:edit]
+  before_action :move_to_index, only: [:edit, :destroy]
   def index
     @items = Item.all.order("created_at DESC")
   end
@@ -38,8 +38,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if current_user.id == @item.user_id
+      item.destroy
+      redirect_to root_path
+    else
+      redirect_to root_path
+    end
   end
 
   private
